# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BlogV2 Backend API
  description: |
    一个现代化的博客后端API，提供用户管理、文章发布、分类标签等功能。
    使用JWT令牌进行身份验证，支持基于角色的权限控制。
  version: 1.0.0
  contact:
    name: API 支持
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发服务器
  - url: https://api.example.com/api/v1
    description: 生产服务器
...
paths:
...
  /posts:
    get:
      tags:
        - 文章
      summary: 获取文章列表（分页）
      description: 获取文章列表，支持分页、筛选和排序。公开访问已发布文章，未发布文章需要管理员权限。
      operationId: getPosts
      parameters:
        - name: page
          in: query
          description: 页码（默认1）
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: 每页数量（默认10）
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: 排序字段（默认createdAt）
          required: false
          schema:
            type: string
            enum: [createdAt, updatedAt, title, views]
            default: createdAt
        - name: sortOrder
          in: query
          description: 排序顺序（默认desc）
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: status
          in: query
          description: 文章状态筛选
          required: false
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
        - name: authorId
          in: query
          description: 按作者筛选
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          description: 按分类筛选
          required: false
          schema:
            type: string
        - name: tagId
          in: query
          description: 按标签筛选
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 成功返回文章列表和分页信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponsePost'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    post:
      tags:
        - 文章
      summary: 创建文章
      description: 创建新的文章。需要认证，编辑或管理员角色。
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostDto'
      responses:
        '201':
          description: 文章创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePost'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非编辑/管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/{id}:
    get:
      tags:
        - 文章
      summary: 根据ID获取文章
      description: 根据文章ID获取文章详细信息。公开访问已发布文章，草稿和归档文章需要管理员权限。
      operationId: getPostById
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: string
      responses:
        '200':
          description: 成功返回文章详细信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePost'
        '404':
          description: 文章不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 没有权限访问此文章
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    patch:
      tags:
        - 文章
      summary: 更新文章
      description: 更新指定文章的信息。需要认证，作者本人或管理员/编辑角色。
      operationId: updatePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostDto'
      responses:
        '200':
          description: 文章更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePost'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非作者/管理员/编辑）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 文章不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

    delete:
      tags:
        - 文章
      summary: 删除文章
      description: 删除指定文章。需要认证，作者本人或管理员角色。
      operationId: deletePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 文章ID
          schema:
            type: string
      responses:
        '200':
          description: 文章删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMessage'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非作者/管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 文章不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/slug/{slug}:
    get:
      tags:
        - 文章
      summary: 根据slug获取文章
      description: 根据文章slug（URL友好格式）获取文章详细信息。公开访问已发布文章，草稿和归档文章需要管理员权限。
      operationId: getPostBySlug
      parameters:
        - name: slug
          in: path
          required: true
          description: 文章slug
          schema:
            type: string
      responses:
        '200':
          description: 成功返回文章详细信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePost'
        '404':
          description: 文章不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 没有权限访问此文章
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /posts/users/{userId}/posts:
    get:
      tags:
        - 文章
      summary: 获取用户的文章列表
      description: 获取指定用户的所有文章。公开访问已发布文章，非公开文章需要管理员权限或本人访问。
      operationId: getUserPosts
      parameters:
        - name: userId
          in: path
          required: true
          description: 用户ID
          schema:
            type: string
        - $ref: '#/components/parameters/pageQuery'
        - $ref: '#/components/parameters/pageSizeQuery'
        - $ref: '#/components/parameters/sortByQuery'
        - $ref: '#/components/parameters/sortOrderQuery'
        - name: status
          in: query
          description: 文章状态筛选
          required: false
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
      responses:
        '200':
          description: 成功返回用户文章列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponsePost'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 没有权限访问用户的非公开文章
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  # ==================== 系统相关接口 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌认证
  parameters:
    pageQuery:
      name: page
      in: query
      description: 页码（默认1）
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSizeQuery:
      name: pageSize
      in: query
      description: 每页数量（默认10）
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    sortByQuery:
      name: sortBy
      in: query
      description: 排序字段
      required: false
      schema:
        type: string
    sortOrderQuery:
      name: sortOrder
      in: query
      description: 排序顺序
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        data:
          type: object
          description: 响应数据，具体类型取决于接口
          nullable: true
        error:
          $ref: '#/components/schemas/ApiError'
          nullable: true
        message:
          type: string
          description: 人类可读的消息
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
      required:
        - success
        - data
        - error
        - message
        - timestamp
    ApiError:
      type: object
      properties:
        code:
          type: string
          description: 错误代码
        message:
          type: string
          description: 错误描述
        details:
          type: object
          description: 错误详情（可选）
          nullable: true
      required:
        - code
        - message
    ApiResponseMessage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
              required:
                - message

    # ===== 用户相关 =====
    User:
      type: object
      properties:
        id:
          type: string
          description: 用户唯一标识符
        username:
          type: string
          description: 用户名
        email:
          type: string
          format: email
          description: 电子邮箱
        role:
          type: string
          enum: [USER, EDITOR, ADMIN]
          description: 用户角色
        avatar:
          type: string
          description: 头像URL
          nullable: true
        bio:
          type: string
          description: 个人简介
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
      required:
        - id
        - username
        - email
        - role
        - createdAt
        - updatedAt
...
    Post:
      type: object
      properties:
        id:
          type: string
          description: 文章唯一标识符
        title:
          type: string
          description: 文章标题
        slug:
          type: string
          description: URL友好标识符
        content:
          type: string
          description: 文章内容（Markdown格式）
        excerpt:
          type: string
          description: 文章摘要
          nullable: true
        featuredImage:
          type: string
          description: 特色图片URL
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          description: 文章状态
        views:
          type: integer
          description: 浏览次数
          default: 0
        authorId:
          type: string
          description: 作者ID
        author:
          $ref: '#/components/schemas/User'
          description: 作者信息
          nullable: true
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
          description: 文章分类
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: 文章标签
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
        publishedAt:
          type: string
          format: date-time
          description: 发布时间
          nullable: true
      required:
        - id
        - title
        - slug
        - content
        - status
        - views
        - authorId
        - createdAt
        - updatedAt
    CreatePostDto:
      type: object
      properties:
        title:
          type: string
          description: 文章标题（1-200个字符）
          minLength: 1
          maxLength: 200
        content:
          type: string
          description: 文章内容（Markdown格式）
        excerpt:
          type: string
          description: 文章摘要（可选）
          nullable: true
        featuredImage:
          type: string
          description: 特色图片URL（可选）
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          description: 文章状态
          default: DRAFT
        categoryIds:
          type: array
          items:
            type: string
          description: 分类ID列表
        tagIds:
          type: array
          items:
            type: string
          description: 标签ID列表
      required:
        - title
        - content
        - categoryIds
    UpdatePostDto:
      type: object
      properties:
        title:
          type: string
          description: 文章标题（1-200个字符）
          minLength: 1
          maxLength: 200
          nullable: true
        content:
          type: string
          description: 文章内容（Markdown格式）
          nullable: true
        excerpt:
          type: string
          description: 文章摘要（可选）
          nullable: true
        featuredImage:
          type: string
          description: 特色图片URL（可选）
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
          description: 文章状态
          nullable: true
        categoryIds:
          type: array
          items:
            type: string
          description: 分类ID列表
          nullable: true
        tagIds:
          type: array
          items:
            type: string
          description: 标签ID列表
          nullable: true
    ApiResponsePost:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Post'
    PaginatedResponsePost:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Post'
            pagination:
              $ref: '#/components/schemas/PaginationInfo'

    # ===== 分类和标签 =====
    Category:
      type: object
      properties:
        id:
          type: string
          description: 分类唯一标识符
        name:
          type: string
          description: 分类名称
        slug:
          type: string
          description: URL友好标识符
        description:
          type: string
          description: 分类描述
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
      required:
        - id
        - name
        - slug
        - createdAt
        - updatedAt
    Tag:
      type: object
      properties:
        id:
          type: string
          description: 标签唯一标识符
        name:
          type: string
          description: 标签名称
        slug:
          type: string
          description: URL友好标识符
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
      required:
        - id
        - name
        - slug
        - createdAt
        - updatedAt

    # ===== 分页信息 =====
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页数量
        totalItems:
          type: integer
          description: 总项目数
        totalPages:
          type: integer
          description: 总页数
        hasNextPage:
          type: boolean
          description: 是否有下一页
        hasPreviousPage:
          type: boolean
          description: 是否有上一页
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
        - hasNextPage
        - hasPreviousPage

    # ===== 健康检查响应 =====
...